+++ src/prelude.rs
@@ -16,7 +16,8 @@ pub use rayon::iter::ParallelIterator;
 pub use shred::AsyncDispatcher;
 
 pub use changeset::ChangeSet;
-pub use storage::{DenseVecStorage, FlaggedStorage, HashMapStorage, InsertedFlag, ModifiedFlag,
-                  NullStorage, ReadStorage, RemovedFlag, Storage, Tracked, VecStorage,
-                  WriteStorage};
+pub use storage::{
+    DenseVecStorage, FlaggedStorage, HashMapStorage, NullStorage, ReadStorage, Storage, Tracked,
+    VecStorage, WriteStorage, ComponentEvent,
+};
 pub use world::{Builder, Component, Entities, Entity, EntityBuilder, LazyUpdate, World};