+++ src/storage/mod.rs
@@ -3,12 +3,14 @@
 pub use self::data::{ReadStorage, WriteStorage};
 pub use self::flagged::FlaggedStorage;
 pub use self::generic::{GenericReadStorage, GenericWriteStorage};
-pub use self::restrict::{ImmutableParallelRestriction, MutableParallelRestriction,
-                         RestrictedStorage, SequentialRestriction};
+pub use self::restrict::{
+    ImmutableParallelRestriction, MutableParallelRestriction, RestrictedStorage,
+    SequentialRestriction,
+};
 #[cfg(feature = "rudy")]
 pub use self::storages::RudyStorage;
 pub use self::storages::{BTreeStorage, DenseVecStorage, HashMapStorage, NullStorage, VecStorage};
-pub use self::track::{InsertedFlag, ModifiedFlag, RemovedFlag, TrackChannels, Tracked};
+pub use self::track::{ComponentEvent, Tracked};
 
 use std;
 use std::marker::PhantomData;
@@ -384,7 +386,8 @@ where
                 }))
             }
         } else {
-            let gen = self.entities
+            let gen = self
+                .entities
                 .alloc
                 .generation(e.id())
                 .unwrap_or(Generation::one());
@@ -445,11 +448,9 @@ where
     }
 }
 
-unsafe impl<'a, T: Component, D> DistinctStorage for Storage<'a, T, D>
-where
-    T::Storage: DistinctStorage,
-{
-}
+unsafe impl<'a, T: Component, D> DistinctStorage for Storage<'a, T, D> where
+    T::Storage: DistinctStorage
+{}
 
 impl<'a, 'e, T, D> Join for &'a Storage<'e, T, D>
 where
@@ -487,8 +488,7 @@ where
     T: Component,
     D: Deref<Target = MaskedStorage<T>>,
     T::Storage: Sync,
-{
-}
+{}
 
 impl<'a, 'e, T, D> Join for &'a mut Storage<'e, T, D>
 where
@@ -518,8 +518,7 @@ where
     T: Component,
     D: DerefMut<Target = MaskedStorage<T>>,
     T::Storage: Sync + DistinctStorage,
-{
-}
+{}
 
 /// Tries to create a default value, returns an `Err` with the name of the storage and/or component
 /// if there's no default.